fix: fix type mismatch on jwt utils

+++ src/main/java/org/mmga/spring/boot/starter/componet/JwtUtils.java
@@ -4,5 +4,6 @@ import java.util.Optional;
 
 public interface JwtUtils {
     String createToken(long uid);
-    Optional<Integer> verifyToken(String token);
+
+    Optional<Long> verifyToken(String token);
 }
